refactor: 拆分 chat_flow 的 Todo 守卫到独立子模块
chat_flow.rs 偏大，将 required Todo Tool 守卫相关逻辑整体迁出到
chat_flow/todo_guard.rs：

- TodoMutationToolKind 及 as_str/matches_executed_tools
- required_todo_tool_kind 及创建/编号引用/最近对象引用/状态屏蔽
  各 helper（detect_create_todo_kind 等私有）
- looks_like_todo_query_text、todo_required_tool_not_called_reply
- 对应 8 项单元测试（位置迁至 chat_flow::todo_guard::tests）

chat_flow.rs 改为通过 todo_guard:: 调用 required_todo_tool_kind /
todo_required_tool_not_called_reply / TodoMutationToolKind；对 super
可见的项标注 pub(super)，私有 helper 保持私有。contains_any 作为
通用 helper 留在父模块，子模块经 use super::contains_any 复用，
不再重复实现；Tool Loop 执行记录、一次重试和本地失败提示机制未改动。

可见性、边界判定、诊断字段和测试结果均不变。
